Made the robot_control page dynamic.
Two major things have happened to this page: 1) it is now rendered
via the mustache engine so that we can pull the current registered
robots from the db and display them in a drop-down (rather than
an input box). 2) The javascript has been abstracted and taught to
deal with the options of the select box rather than the text field.